requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
easyocr==1.7.2
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fsspec==2024.10.0
idna==3.10
imageio==2.36.1
itsdangerous==2.2.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.0
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pyclipper==1.3.0.post6
python-bidi==0.6.3
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.0
scipy==1.14.1
shapely==2.0.6
sympy==1.13.3
tifffile==2024.12.12
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
